Edited database construction and credentials path
Fixed typos in database.py and fixed issue with creating unique constraints. Revised the portion to retrieve the file 'pipeline.conf assuming the file is in the same directory as the pipeline_temporary.conf file (which is optional to use). Also edited the README file to include a bit about setting up airflow.